fix(build): switch back to babel plugins and remove sass preprocessor from dependencies
To overcome dependency issues with babel 6 and 7 plugins we will rely on babel 7 plugins for future development. We will also tend to use only babel own plugins in the future to be always on the latest track of the babel developments. Because we don't need an html frontend for the plugin we will remove all dependencies regarding ui frameworks. We will also remove sass preprocessors dependencies, because we don#t need the anymore.