feat(src): add first version of safari app extension
The first version of the safari app extension imarex will contain two content scripts:

script.js -> triggers the iom.c function
imarex.js -> hosts the iam.js source code modified to work in safari app extension

This is an WIP version of the imarex extension for safari. There are no build options for this version yet.
We plant to implement a build process which will be running only on mac for the next releases.