chore(build): add build process for safari extension
This change makes it possible to build the safari extension on Mac computer with installed xcode cli build tools. The build process has three stages. First stage will copy all source files to the .temp directory. In the second stage the webpack compiler kicks in and will compile the source code to es2015 compliant code and will swap global constants with the configured values. The last and third stage will exec the xcode cli build tools to compile the whole app with the safari extension to an Apple app.