fix: activate tracking only when panel identifier was transmitted
This will fix an issue where the tracking is activated too early in the
registration process. It will be now activated when the panel identifier
was transmitted to the extension and not before that event. This fix
will also deactivate the tracking when the provider or the registration
it self was removed from extension state.

Fixes #20